feat(api): chat polish, throttling, OpenAPI, Jest 29 stack
- Split chat into room/messages services; WS validation, rate limit, UUID params
- Shared HTTP bootstrap (CORS, trust proxy, ValidationPipe) for main and e2e
- Throttle config module; ChatSocketThrottleService; user-scoped throttler guard
- OpenAPI: /rooms paths and schemas; user search trgm migration
- tsconfig.e2e.json; typecheck without redundant prisma generate; Jest 29.7 + ts-jest aligned

Made-with: Cursor
