+++ apps/api/.env.example
@@ -11,6 +11,13 @@ JWT_REFRESH_EXPIRES_DAYS=7
 # CORS: frontend origin(s), comma-separated. Required for cookie auth across origins.
 CORS_ORIGIN=http://localhost:5173,http://localhost:3000
 
+# Reverse proxy: set to true or 1 so Express trusts X-Forwarded-* (correct IP for rate limits).
+# TRUST_PROXY=true
+
+# Global rate limit defaults (GET /users/search uses ThrottlerUserGuard). Milliseconds / max hits per window.
+# THROTTLE_TTL_MS=60000
+# THROTTLE_LIMIT=30
+
 # Cookie SameSite: lax (same-site dev) | none (cross-site SPA + API, needs HTTPS in prod)
 COOKIE_SAME_SITE=lax
