+++ apps/api/jest.config.json
@@ -7,7 +7,7 @@
     "^.+\\.(t|j)s$": [
       "ts-jest",
       {
-        "tsconfig": "tsconfig.json"
+        "tsconfig": "tsconfig.e2e.json"
       }
     ]
   },
