+++ apps/api/package.json
@@ -9,7 +9,7 @@
     "start": "node dist/main.js",
     "lint": "eslint .",
     "test": "jest --config jest.config.json --runInBand",
-    "typecheck": "prisma generate && tsc --noEmit -p tsconfig.json",
+    "typecheck": "tsc --noEmit -p tsconfig.json && tsc --noEmit -p tsconfig.e2e.json",
     "prisma:generate": "prisma generate",
     "prisma:migrate:dev": "prisma migrate dev"
   },
@@ -51,7 +51,7 @@
     "prisma": "^6.19.2",
     "socket.io-client": "^4.8.3",
     "supertest": "^7.2.2",
-    "ts-jest": "^29.4.6",
+    "ts-jest": "^29.4.9",
     "typescript": "~5.7.2"
   }
 }
