+++ apps/api/src/app.module.ts
@@ -1,6 +1,7 @@
 import { Module } from "@nestjs/common";
 import { ThrottlerModule } from "@nestjs/throttler";
 import { AppController } from "./app.controller";
+import { throttlerOptionsFromEnv } from "./config/throttle.config";
 import { AuthModule } from "./modules/auth/auth.module";
 import { ChatModule } from "./modules/chat/chat.module";
 import { PrismaModule } from "./modules/prisma/prisma.module";
@@ -9,12 +10,9 @@ import { UsersModule } from "./modules/users/users.module";
 @Module({
   imports: [
     PrismaModule,
-    ThrottlerModule.forRoot([
-      {
-        ttl: 60_000,
-        limit: 30,
-      },
-    ]),
+    ThrottlerModule.forRootAsync({
+      useFactory: () => throttlerOptionsFromEnv(),
+    }),
     AuthModule,
     ChatModule,
     UsersModule,
