+++ apps/api/src/config/cors.config.ts
@@ -1,29 +1,41 @@
 import type { CorsOptions } from "@nestjs/common/interfaces/external/cors-options.interface";
 
-export function createCorsOptions(): CorsOptions {
-  const configuredCorsOrigins =
+function getEffectiveCorsOrigins(): string[] {
+  const configured =
     process.env.CORS_ORIGIN?.split(",")
       .map((s) => s.trim())
       .filter(Boolean) ?? [];
-  const isDev = process.env.NODE_ENV !== "production";
-  const fallbackDevOrigins = ["http://localhost:5173", "http://localhost:3000"];
-  const corsOrigins =
-    configuredCorsOrigins.length > 0
-      ? configuredCorsOrigins
-      : isDev
-        ? fallbackDevOrigins
-        : [];
+  if (configured.length > 0) {
+    return configured;
+  }
+  return process.env.NODE_ENV !== "production"
+    ? ["http://localhost:5173", "http://localhost:3000"]
+    : [];
+}
 
+export function isCorsOriginAllowed(origin: string | undefined): boolean {
+  if (!origin) {
+    return true;
+  }
+  const corsOrigins = getEffectiveCorsOrigins();
+  return corsOrigins.length > 0 && corsOrigins.includes(origin);
+}
+
+function corsOriginResolver(
+  origin: string | undefined,
+  callback: (err: Error | null, allow?: boolean) => void,
+): void {
+  callback(null, isCorsOriginAllowed(origin));
+}
+
+export function createCorsOptions(): CorsOptions {
   return {
-    origin: (origin, callback) => {
-      if (!origin) {
-        return callback(null, true);
-      }
-      if (corsOrigins.includes(origin)) {
-        return callback(null, true);
-      }
-      callback(null, false);
-    },
+    origin: corsOriginResolver,
     credentials: true,
   };
 }
+
+export const socketIoCors = {
+  origin: corsOriginResolver,
+  credentials: true,
+};
