+++ apps/api/src/main.ts
@@ -1,23 +1,14 @@
 import "dotenv/config";
-import { ValidationPipe } from "@nestjs/common";
 import { NestFactory } from "@nestjs/core";
-import cookieParser from "cookie-parser";
-import { createCorsOptions } from "./config/cors.config";
+import { NestExpressApplication } from "@nestjs/platform-express";
+import { applyHttpAppSetup, applyTrustProxy } from "./bootstrap/http-app-setup";
 import { AppModule } from "./app.module";
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
+  const app = await NestFactory.create<NestExpressApplication>(AppModule);
 
-  app.use(cookieParser());
-  app.enableCors(createCorsOptions());
-
-  app.useGlobalPipes(
-    new ValidationPipe({
-      whitelist: true,
-      forbidNonWhitelisted: true,
-      transform: true,
-    }),
-  );
+  applyTrustProxy(app);
+  applyHttpAppSetup(app);
 
   const port = process.env.PORT ?? "3000";
   await app.listen(port);
