+++ apps/api/src/modules/chat/chat.module.ts
@@ -1,13 +1,24 @@
 import { Module } from "@nestjs/common";
-import { JwtAuthGuard } from "../auth/guards/jwt-auth.guard";
+import { AuthModule } from "../auth/auth.module";
 import { ChatController } from "./chat.controller";
 import { ChatGateway } from "./chat.gateway";
+import { ChatMessagesService } from "./chat-messages.service";
 import { ChatPresenceService } from "./chat-presence.service";
+import { ChatRoomService } from "./chat-room.service";
+import { ChatSocketThrottleService } from "./chat-socket-throttle.service";
 import { ChatService } from "./chat.service";
 
 @Module({
+  imports: [AuthModule],
   controllers: [ChatController],
-  providers: [ChatService, ChatPresenceService, JwtAuthGuard, ChatGateway],
+  providers: [
+    ChatRoomService,
+    ChatMessagesService,
+    ChatService,
+    ChatPresenceService,
+    ChatSocketThrottleService,
+    ChatGateway,
+  ],
   exports: [ChatService],
 })
 export class ChatModule {}
