+++ apps/api/src/modules/users/users.controller.ts
@@ -1,10 +1,10 @@
 import { Controller, Get, Query, UseGuards } from "@nestjs/common";
-import { ThrottlerGuard } from "@nestjs/throttler";
 import { CurrentUser } from "../auth/decorators/current-user.decorator";
 import { JwtAuthGuard } from "../auth/guards/jwt-auth.guard";
 import { JwtPayload } from "../auth/types/jwt-payload.type";
 import { SearchUsersQueryDto } from "./dto/search-users.query.dto";
 import { UserSearchItemResponse } from "./types/user-search-response.type";
+import { ThrottlerUserGuard } from "./throttler-user.guard";
 import { UsersService } from "./users.service";
 
 @UseGuards(JwtAuthGuard)
@@ -13,7 +13,7 @@ export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Get("search")
-  @UseGuards(ThrottlerGuard)
+  @UseGuards(ThrottlerUserGuard)
   searchUsers(
     @CurrentUser() user: JwtPayload,
     @Query() query: SearchUsersQueryDto,
