+++ apps/api/src/modules/users/users.module.ts
@@ -1,10 +1,12 @@
 import { Module } from "@nestjs/common";
-import { JwtAuthGuard } from "../auth/guards/jwt-auth.guard";
+import { AuthModule } from "../auth/auth.module";
+import { ThrottlerUserGuard } from "./throttler-user.guard";
 import { UsersController } from "./users.controller";
 import { UsersService } from "./users.service";
 
 @Module({
+  imports: [AuthModule],
   controllers: [UsersController],
-  providers: [UsersService, JwtAuthGuard],
+  providers: [UsersService, ThrottlerUserGuard],
 })
 export class UsersModule {}
